wxauto-restful-api/pyproject.toml
[project]
name = "wxauto-api"
version = "2.0.0"
description = "微信自动化API服务，统一使用 wxautox4"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "wxautox>=39.1.29",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.20",
    "websockets>=12.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
